UF-14187: Changed access mode to enums not visible in requests
- Some of the enums used in both read and write scenarios have the
access mode set to different values, but this creates a problem as the
generated swagger will use the first available value for attribute
(either READ_ONLY or WRITE_ONLY). If the value is set to READ_ONLY, the
attribute will not show up in the model for the POST and PATCH resources
using the enum reference. Fix is to remove access mode for enums used in
both read and write scenarios.